C. Wood [#4]: Separating logic into new lib.rs file.
